Raporlar/İREM_BSG/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
argparse-addons==0.12.0
attrs==25.4.0
bitstruct==8.21.0
cantools==40.7.1
crccheck==1.3.1
diskcache==5.6.3
frozenlist==1.8.0
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.7.0
ocpp==2.1.0
packaging==25.0
propcache==0.4.1
python-can==4.6.1
referencing==0.37.0
rpds-py==0.28.0
textparser==0.24.0
typing_extensions==4.15.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
